# docs/swagger.yaml
basePath: /v1/user-services
definitions:
  apperror.Code:
    enum:
    - BAD_REQUEST
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - METHOD_NOT_ALLOWED
    - NOT_ACCEPTABLE
    - REQUEST_TIMEOUT
    - CONFLICT
    - GONE
    - PRECONDITION_FAILED
    - PAYLOAD_TOO_LARGE
    - UNSUPPORTED_MEDIA_TYPE
    - VALIDATION_ERROR
    - UNPROCESSABLE_ENTITY
    - LOCKED
    - FAILED_DEPENDENCY
    - PRECONDITION_REQUIRED
    - TOO_MANY_REQUESTS
    - INTERNAL_ERROR
    - NOT_IMPLEMENTED
    - BAD_GATEWAY
    - SERVICE_UNAVAILABLE
    - TIMEOUT
    - CACHE_MISS
    - CACHE_ERROR
    - MESSAGE_BUS_ERROR
    type: string
    x-enum-varnames:
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeMethodNotAllowed
    - CodeNotAcceptable
    - CodeRequestTimeout
    - CodeConflict
    - CodeGone
    - CodePrecondFailed
    - CodePayloadTooLarge
    - CodeUnsupportedMedia
    - CodeValidation
    - CodeUnprocessable
    - CodeLocked
    - CodeFailedDependency
    - CodePrecondRequired
    - CodeTooManyRequest
    - CodeInternal
    - CodeNotImplemented
    - CodeBadGateway
    - CodeServiceUnavail
    - CodeTimeout
    - CodeCacheMiss
    - CodeCacheError
    - CodeMessageBus
...
  github_com_modami_user-service_internal_dto.UpdateAvatarRequest:
    properties:
      avatar_url:
        type: string
    required:
    - avatar_url
    type: object
  github_com_modami_user-service_internal_dto.UpdateCoverRequest:
    properties:
      cover_url:
        type: string
    required:
    - cover_url
    type: object
  github_com_modami_user-service_internal_dto.UpdateProfileRequest:
    properties:
      bio:
        maxLength: 500
        type: string
      date_of_birth:
        type: string
      full_name:
        maxLength: 255
        minLength: 1
        type: string
      gender:
        enum:
        - male
        - female
        - other
        - undisclosed
        type: string
      phone:
        type: string
    type: object
...
  github_com_modami_user-service_internal_dto.UserProfileResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      cover_url:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      follower_count:
        type: integer
      following_count:
        type: integer
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      phone:
        type: string
      role:
        type: string
      status:
        type: string
      trust_score:
        type: number
    type: object
  response.ErrorBody:
    properties:
      code:
        $ref: '#/definitions/apperror.Code'
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        type: string
    type: object
  response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/response.Pagination'
      request_id:
        type: string
      timestamp:
        type: integer
    type: object
  response.Pagination:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page:
        type: integer
      per_page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8086
info:
  contact: {}
  description: User service for the Modami marketplace platform.
  title: Modami User Service API
  version: "1.0"
paths:
...
  /users/{id}:
    get:
      description: Returns a user's public profile by ID
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get user profile
      tags:
      - Users
...
  /users/me:
    delete:
      description: Deactivates the authenticated user's account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Deactivate account
      tags:
      - Users
    get:
      description: Returns the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's profile fields
      parameters:
      - description: Profile fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - Users
...
  /users/me/avatar:
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's avatar URL
      parameters:
      - description: Avatar URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.UpdateAvatarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update avatar
      tags:
      - Users
  /users/me/cover:
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's cover image URL
      parameters:
      - description: Cover URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.UpdateCoverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update cover image
      tags:
      - Users
...
  /users/search:
    get:
      description: Search users by name or email
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Results per page (max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Search users
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"